+++ plugins/engine/3d/engine.cpp
@@ -1595,8 +1595,8 @@ void csEngine::ShineLights (iRegion *region, iProgressMeter *meter)
   }
   if (failed > 0)
   {
-    Warn ("Couldn't load cached lighting for %lu object(s):",
-	  (unsigned long)failed);
+    Warn ("Couldn't load cached lighting for %zu object(s):",
+	  failed);
     size_t i;
     for (i = 0 ; i < failed_meshes.Length () ; i++)
     {
